#36 foe script
renamed folders
