+++ src/game_objects/projectiles/projectile.py
@@ -1,22 +1,24 @@
-from src.game_object.dynamic import Dynamic
+import pygame
+
+from src.game_objects.dynamic import Dynamic
+from src.game_objects.foes.foe import Foe
 
 
-class Scenic(Dynamic):
+class Projectile(Dynamic):
     
     def __init__(self, game, *, pos, depth, image):
         super().__init__(game, pos=pos, depth=depth, image=image)
-    
+        
     def decayable(self):
         """
         Overriding decayable in GameObject
         :return: bool
         """
-        return self.rect.right < 0
-    
+        active_zone = self.game.surface.get_rect()
+        return not active_zone.colliderect(self.rect)
+
     def draw(self):
         super().draw()
-    
+
     def update(self):
         super().update()
-        
-        self.x += -self.game.pan_speed
