+++ src/game_objects/scenic.py
@@ -1,24 +1,22 @@
-import pygame
-
-from src.game_object.dynamic import Dynamic
-from src.game_object.foe.foe import Foe
+from src.game_objects.dynamic import Dynamic
 
 
-class Projectile(Dynamic):
+class Scenic(Dynamic):
     
     def __init__(self, game, *, pos, depth, image):
         super().__init__(game, pos=pos, depth=depth, image=image)
-        
+    
     def decayable(self):
         """
         Overriding decayable in GameObject
         :return: bool
         """
-        active_zone = self.game.surface.get_rect()
-        return not active_zone.colliderect(self.rect)
-
+        return self.rect.right < 0
+    
     def draw(self):
         super().draw()
-
+    
     def update(self):
         super().update()
+        
+        self.x += -self.game.pan_speed
