+++ src/game_objects/wall.py
@@ -1,7 +1,7 @@
 import pygame
 
 from src.const import *
-from src.game_object.scenic import Scenic
+from src.game_objects.scenic import Scenic
 
 
 class Wall(Scenic):
